Make `dev server respond with index.html to all unrecognized requests
As of today, if anyone uses this template to create an app with some kind of routing library (e.g. `svelte-routing`) and they reload in any url but the root, the server won't respond.

I think this is not a good default for this template. Passing `--single index.html` will serve the index.html instead, which I think is nicer default.