Proposing that --single to npm run dev
I propose that the flag `--single` be added to the `npm run dev` script. This will allow fallback routing to work with the development server. 

This is important as new folks use this template to get started. When they add client side routing and they refresh on a client side route, the current behavior will result in a 404. By adding `--single` this will "just work"

This is documented in the readme here, but I believe that this should be the default.