Added host flag in npm run start
Added --host flag in npm run start to start the server on the wifi network.

A very common question in our community is how to start a server on a Wi-Fi network. I suggest adding this flag to the template, as it is a useful feature that will not interfere with the development process.